Create BookmarkRepository class
