+++ phpstan-baseline.neon
@@ -121,44 +121,34 @@ parameters:
 			path: src/Advisory/Advisor.php
 
 		-
-			message: "#^Cannot cast mixed to int\\.$#"
-			count: 1
-			path: src/Bookmark.php
-
-		-
 			message: "#^Construct empty\\(\\) is not allowed\\. Use more strict comparison\\.$#"
 			count: 1
-			path: src/Bookmark.php
+			path: src/Bookmarks/Bookmark.php
 
 		-
 			message: "#^Parameter \\#1 \\$str of method PhpMyAdmin\\\\DatabaseInterface\\:\\:escapeString\\(\\) expects string, mixed given\\.$#"
 			count: 1
-			path: src/Bookmark.php
+			path: src/Bookmarks/Bookmark.php
 
 		-
-			message: "#^Parameter \\#2 \\$row of static method PhpMyAdmin\\\\Bookmark\\:\\:createFromRow\\(\\) expects array, mixed given\\.$#"
-			count: 1
-			path: src/Bookmark.php
+			message: "#^Parameter \\#2 \\$row of static method PhpMyAdmin\\\\Bookmarks\\\\BookmarkRepository\\:\\:createFromRow\\(\\) expects array\\<string\\>, array given\\.$#"
+			count: 2
+			path: src/Bookmarks/BookmarkRepository.php
 
 		-
-			message: "#^Property PhpMyAdmin\\\\Bookmark\\:\\:\\$currentUser \\(string\\) does not accept mixed\\.$#"
+			message: "#^Parameter \\#2 \\$row of static method PhpMyAdmin\\\\Bookmarks\\\\BookmarkRepository\\:\\:createFromRow\\(\\) expects array\\<string\\>, mixed given\\.$#"
 			count: 1
-			path: src/Bookmark.php
+			path: src/Bookmarks/BookmarkRepository.php
 
 		-
-			message: "#^Property PhpMyAdmin\\\\Bookmark\\:\\:\\$database \\(string\\) does not accept mixed\\.$#"
+			message: "#^Property PhpMyAdmin\\\\Bookmarks\\\\BookmarkRepository\\:\\:\\$dbi is never read, only written\\.$#"
 			count: 1
-			path: src/Bookmark.php
+			path: src/Bookmarks/BookmarkRepository.php
 
 		-
-			message: "#^Property PhpMyAdmin\\\\Bookmark\\:\\:\\$label \\(string\\) does not accept mixed\\.$#"
+			message: "#^Property PhpMyAdmin\\\\Bookmarks\\\\BookmarkRepository\\:\\:\\$relation is never read, only written\\.$#"
 			count: 1
-			path: src/Bookmark.php
-
-		-
-			message: "#^Property PhpMyAdmin\\\\Bookmark\\:\\:\\$query \\(string\\) does not accept mixed\\.$#"
-			count: 1
-			path: src/Bookmark.php
+			path: src/Bookmarks/BookmarkRepository.php
 
 		-
 			message: "#^Cannot access offset mixed on mixed\\.$#"
