+++ psalm-baseline.xml
@@ -39,30 +39,43 @@
       <code>$value</code>
     </MixedAssignment>
   </file>
-  <file src="src/Bookmark.php">
+  <file src="src/Bookmarks/Bookmark.php">
     <DeprecatedMethod>
-      <code>Config::getInstance()</code>
-      <code>Config::getInstance()</code>
-      <code>Config::getInstance()</code>
       <code>escapeString</code>
     </DeprecatedMethod>
     <MixedArgument>
-      <code>$row</code>
       <code>$variables[$i]</code>
     </MixedArgument>
-    <MixedAssignment>
-      <code><![CDATA[$bookmark->currentUser]]></code>
-      <code><![CDATA[$bookmark->database]]></code>
-      <code><![CDATA[$bookmark->label]]></code>
-      <code><![CDATA[$bookmark->query]]></code>
-      <code>$row</code>
-    </MixedAssignment>
     <PossiblyUnusedMethod>
       <code>getDatabase</code>
     </PossiblyUnusedMethod>
     <PossiblyUnusedReturnValue>
       <code>bool</code>
     </PossiblyUnusedReturnValue>
+  </file>
+  <file src="src/Bookmarks/BookmarkRepository.php">
+    <DeprecatedMethod>
+      <code>Config::getInstance()</code>
+      <code>Config::getInstance()</code>
+      <code>Config::getInstance()</code>
+    </DeprecatedMethod>
+    <MixedArgument>
+      <code>$row</code>
+    </MixedArgument>
+    <MixedArgumentTypeCoercion>
+      <code>$result</code>
+      <code>$result</code>
+    </MixedArgumentTypeCoercion>
+    <MixedAssignment>
+      <code>$row</code>
+    </MixedAssignment>
+    <PossiblyUnusedMethod>
+      <code>__construct</code>
+    </PossiblyUnusedMethod>
+    <UnusedProperty>
+      <code>$dbi</code>
+      <code>$relation</code>
+    </UnusedProperty>
   </file>
   <file src="src/BrowseForeigners.php">
     <InvalidArrayOffset>
